feat: Introduce main application class for Users service
A fresh file named UsersServiceApplication.java has been added to the project. This file serves as the anchoring class for bootstrapping the User Service application. It carries the @SpringBootApplication annotation for vital configurations and autoconfiguration enablement. The incorporated main method marks the entry point of the application and marshals the application using SpringApplication.run() method.