feat (docker-compose.yml): initialize services in containers
A docker-compose.yml file has been added. It initializes a mysql8 service and a 'users' service, each in their own container. The 'users' service depends on the mysql8 service and both are part of an 'e-commerce-backend-java' network. The docker-compose file also specifies volumes and environment variables required for the services.